B3/eisen_inventory/requirements.txt
Django>=4.2,<5.3
djangorestframework>=3.15
openpyxl>=3.1
psycopg2-binary>=2.9 ; platform_system != 'Windows'
# On Windows prefer sqlite or install: pip install psycopg2-binary (ensure Visual C++ Build Tools)
django-cors-headers>=4.4
python-dotenv>=1.0
whitenoise>=6.7
dj-database-url>=2.3
